image/test/mochitest/test_bug1180105.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1180105
-->
<head>
  <title>Test for Bug 1180105</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="initializeOnload()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1180105">Mozilla Bug 1180105</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

SimpleTest.waitForExplicitFinish();

const WAITER_URL = "bug1180105-waiter.sjs";

function initializeOnload() {
  var firstimg = document.createElement('img');
  firstimg.src = "bug1180105.sjs";
  document.getElementById('content').appendChild(firstimg);

  waitForFinish();
}

function waitForFinish() {
  var loader = document.getElementById("loader");
  loader.src = WAITER_URL;
  loader.onload = function() {
    var img = document.getElementsByTagName('img')[0];
    ok(img.width > 0, "Image should be loaded by now");
    SimpleTest.finish();
  };
}

</script>
</pre>
<div id="content">>
<iframe id="loader"></iframe>
</div>
</body>
</html>
